requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   starlette
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via click
distro==1.9.0
    # via google-genai
fastapi==0.128.0
    # via -r requirements.in
google-auth[requests]==2.47.0
    # via google-genai
google-genai==1.57.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
    #   google-genai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
python-dotenv==1.2.1
    # via pydantic-settings
python-multipart==0.0.21
    # via -r requirements.in
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via google-genai
starlette==0.50.0
    # via fastapi
tenacity==9.1.2
    # via google-genai
typing-extensions==4.15.0
    # via
    #   fastapi
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
uvicorn==0.40.0
    # via -r requirements.in
websockets==15.0.1
    # via google-genai
